LayoutTests/fast/css/scale-invalidate-if-disabled.html

<!DOCTYPE html><!-- webkit-test-runner [ CSSIndividualTransformPropertiesEnabled=false ] -->
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/js-test.js"></script>
</head>
<body>
<div id="target" style="display: none"></div>
<script>
description("Tests that scale is not exposed when the feature is disabled");

shouldBeFalse("'scale' in document.documentElement.style");
shouldBeFalse("'scale' in getComputedStyle(document.documentElement)");
shouldBeTrue("getComputedStyle(document.getElementById('target')).getPropertyValue('scale') == ''");
shouldBeFalse("CSS.supports('scale: contain')");
shouldBeFalse("CSS.supports('scale: inherit')");
shouldBeFalse("CSS.supports('scale', 'inherit')");

</script>
</body>
</html>
