LayoutTests/fast/css/scale-invalidate-if-disabled-expected.txt

Tests that scale is not exposed when the feature is disabled

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'scale' in document.documentElement.style is false
PASS 'scale' in getComputedStyle(document.documentElement) is false
PASS getComputedStyle(document.getElementById('target')).getPropertyValue('scale') == '' is true
PASS CSS.supports('scale: contain') is false
PASS CSS.supports('scale: inherit') is false
PASS CSS.supports('scale', 'inherit') is false
PASS successfullyParsed is true

TEST COMPLETE
